+++ bzfs_tests/itest/test_integrations.py
@@ -29,6 +29,9 @@ from bzfs_tests.itest.ibase import (
 from bzfs_tests.itest.test_incremental_send_steps import (
     IncrementalSendStepsTestCase,
 )
+from bzfs_tests.itest.test_lima_ubuntu_24_04_sh import (
+    TestLimaUbuntuScript,
+)
 from bzfs_tests.itest.test_local import (
     LocalTestCase,
 )
@@ -44,15 +47,12 @@ from bzfs_tests.itest.test_ssh import (
     TestSSHLatency,
     TestSSHMasterIntermittentFailure,
 )
-from bzfs_tests.itest.test_test_ubuntu_24_04_lima_sh import (
-    TestUbuntuLimaScript,
-)
 
 
 def suite() -> unittest.TestSuite:
     ttype = AbstractTestCase()
     suite = unittest.TestSuite()
-    suite.addTest(unittest.TestLoader().loadTestsFromTestCase(TestUbuntuLimaScript))
+    suite.addTest(unittest.TestLoader().loadTestsFromTestCase(TestLimaUbuntuScript))
     if not (ttype.is_smoke_test or ttype.is_functional_test or ttype.is_adhoc_test):
         suite.addTest(ParametrizedTestCase.parametrize(IncrementalSendStepsTestCase, {"verbose": True}))
         suite.addTest(ParametrizedTestCase.parametrize(TestSSHLatency))